+++ openmls/src/group/mls_group/errors.rs
@@ -181,6 +181,7 @@ pub enum ProcessMessageError<StorageError> {
 }
 
 /// Create message error
+#[cfg(not(feature = "virtual-clients-draft"))]
 #[derive(Error, Debug, PartialEq, Clone)]
 pub enum CreateMessageError {
     /// See [`LibraryError`] for more details.
@@ -551,4 +552,58 @@ pub enum RemoveProposalError<StorageError> {
     /// Error erasing proposal from storage.
     #[error("error writing proposal to storage")]
     Storage(StorageError),
+}
+
+#[cfg(feature = "virtual-clients-draft")]
+pub use virtual_clients_draft::*;
+
+#[cfg(feature = "virtual-clients-draft")]
+mod virtual_clients_draft {
+    use thiserror::Error;
+
+    use super::MlsGroupStateError;
+    use crate::error::LibraryError;
+    use crate::framing::MessageEncryptionError;
+    use crate::tree::secret_tree::SecretTreeError;
+
+    /// Create message error
+    #[derive(Error, Debug, PartialEq, Clone)]
+    pub enum CreateMessageError<StorageError> {
+        /// See [`LibraryError`] for more details.
+        #[error(transparent)]
+        LibraryError(#[from] LibraryError),
+        /// See [`MlsGroupStateError`] for more details.
+        #[error(transparent)]
+        GroupStateError(#[from] MlsGroupStateError),
+        /// See [`MessageEncryptionError`] for more details.
+        #[error(transparent)]
+        MessageEncryptionError(#[from] MessageEncryptionError<StorageError>),
+        /// Error writing to storage.
+        #[error("Error writing to storage: {0}")]
+        StorageError(StorageError),
+    }
+
+    /// Confirm message error
+    #[derive(Error, Debug, PartialEq, Clone)]
+    pub enum ConfirmMessageError<StorageError> {
+        /// See [`SecretTreeError`] for more details.
+        #[error(transparent)]
+        SecretTreeError(#[from] SecretTreeError),
+        /// Error writing to storage.
+        #[error("Error writing to storage: {0}")]
+        StorageError(StorageError),
+    }
+
+    impl<StorageError> From<ConfirmMessageError<StorageError>> for CreateMessageError<StorageError> {
+        fn from(err: ConfirmMessageError<StorageError>) -> Self {
+            match err {
+                ConfirmMessageError::SecretTreeError(e) => {
+                    CreateMessageError::MessageEncryptionError(
+                        MessageEncryptionError::SecretTreeError(e),
+                    )
+                }
+                ConfirmMessageError::StorageError(e) => CreateMessageError::StorageError(e),
+            }
+        }
+    }
 }